backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy>=2.0
alembic
psycopg[binary]
passlib[argon2]
pyjwt[crypto]
google-auth
requests
email-validator
pytest
httpx